+++ src/globalInfo/index.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2021, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export * from './globalInfoConfig';
+export * from './sfdxDataHandler';
+export * from './types';
+export { OrgAccessor } from './accessors/orgAccessor';
+export { AliasAccessor } from './accessors/aliasAccessor';
+export { TokenAccessor } from './accessors/tokenAccessor';
