+++ test/unit/org/userTest.ts
@@ -8,7 +8,7 @@ import { stubMethod } from '@salesforce/ts-sinon';
 import { AnyJson } from '@salesforce/ts-types';
 import { expect, assert } from 'chai';
 import { AuthInfo } from '../../../src/org/authInfo';
-import { OrgAccessor } from '../../../src/config/globalInfoConfig';
+import { OrgAccessor } from '../../../src/globalInfo';
 import { Connection } from '../../../src/org/connection';
 import { Org } from '../../../src/org/org';
 import { PermissionSetAssignment } from '../../../src/org/permissionSetAssignment';
